Add stub object file for `main()`
Unit tests and other linkage requires that the ardopcf executable's
main function be separate from the rest of the code. That way, we
can build programs that contain all parts of ardopcf except for
`main()`.

Since the Linux and Windows executables are so very different, and
each depends on different globals, we can't easily `ifdef` this
out. Instead, rename these existing main functions to
`platform_main()` and invoke them from within a new file.