Make ASCII compression more robust
The `ASCIIto6Bit()` compression function requires, but does not
check, input in range ASCII 32 to ASCII 95. It also behaves poorly
when the string contains NULs because it is not quite 8 characters
long.

Instead of relying on the caller to get it right, fix the input
to the extent possible. The input is padded out with spaces and
uppercased if necessary.

One might be tempted to use `toupper(3)` for this, but this is
actually wrong. `toupper()` observes the current locale, if one
is enabled with `setlocale()`. Some locales have unexpected
uppercase behavior across the ASCII range [1]. If we want a
strictly-ASCII uppercase, we need to do it ourselves.

References

[1]: https://github.com/util-linux/util-linux/commit/64a89adaa16461bd58730c1408b30aa7244b4f66