Add cmocka-based tests for callsign compression
Add Makefile rules for a cmocka-based test suite. Each `.c` file
in `tests/ardop/` compiles to its own test-case executable. The cmocka
libraries must be installed on the host, such as:

    sudo apt install libcmocka-dev

The `make buildtest` target compiles all test-cases.

The `make test` target compiles **and** executes all test-cases.

The call sign compress/decompress logic is targeted for the
initial round of tests since it is easy to understand and has
the following observed defects:

* Use of unbounded string functions `strcpy()` and `strcat()`

* For compression, that the input buffer has a minimum
  length that is potentially greater than its size. This
  restriction is not documented and makes it unsafe to use
  with string constants.

* For decompression, the output is variable-length without
  accepting a buffer size. This can cause a crash bug.

Testing reveals that these functions additionally do not work
with seven-character callsigns, which the spec requires we
support.

Co-authored-by: pflarue <131492283+pflarue@users.noreply.github.com>