Make txt2c part of the build
The webgui presently relies on converting the HTML and JavaScript
files into C sources ahead of time. The generated .c sources are
committed to the repo. This is neither merge-friendly nor
necessary.

Instead, just include txt2c in the build as a prerequisite. Add
rules for making the web files into .c files. Implicit rules take
care of the rest.

Make cleans up the generated sources for us.

Since the Windows delete command is vastly different from the
Linux delete command, just make a separate ifeq/else block for it.