Move ARDOPC/ → src/ and move Makefile to root
The `ARDOPC/` directory is no longer necessary and is removed.
C sources are reorganized under a new top-level `src/`
directory. The Makefile is relocated to the repository root for
easy access.

Remove the `vpath` make definitions. This eliminates potential
filename conflicts. Instead, each file should be listed by its
full path relative to the Makefile directory. (i.e., the repo
root.)

The `src/` directory is now the root of the `-I` include path
for *all* ARDOPC sources. Include paths should be relative to
either:

* `src/`; OR
* `lib/`, which contains subprojects and third-party sources

All includes in `ARDOPC/` should now be relative to one of
those two places. This avoids weird relative paths like:

    #include "../lib/something/else.h"   /* deprecated */

and also minimizes the number of `-I` include paths we have to set.

Includes within `src/` are rewritten to use the full path in
every include, like:

    #include "common/version.h"

instead of

    #include "version.h"

This is less ambiguous and more robust.

The lib sources are not affected by any of these changes and are
untouched.

We split sources that are platform-specific out into their own
subdirectory. Additional suggestions for subdirectory
organization are welcome.

A very pleasant side-effect is that the top-level directory is
much less cluttered after a build. The executable ends up in the
repo root.

Eventually there will be a `test/` directory for unit tests.