Always do rs_correct() before checking CRC of data frames
Previously, if CRC showed a data frame to be valid, rs_correct()
was not done.  While unlikely, a case was encountered in which CRC
showed the data to be valid, even through it was not.  In that
case rs_correct() was able to successfully correct it.  So, always
do rs_correct() before CheckCRC16FrameType().  The increased
computational burden is relatively low, and it reduces the
likelyhood of an undetected error.