linux: signal handler safety
On Linux, the signal handlers call `printf()`, which is not
async-signal-safe. See [`man 7 signal-safety`](1). printf must not
be used in signal handlers.

Instead of directly printing, we record the signal ID that caused
termination and print it later. Unfortunately, `strsignal()` is
not particularly safe, and `sigabbrev_np()` is not portable. We
instead write our own implementation,
`PlatformSignalAbbreviation()`.

This fixes UB that can occur from a user interrupt like `^C`.

[1]: https://man7.org/linux/man-pages/man7/signal-safety.7.html