locator: add object for Maidenhead Grid Square
Previous versions of ARDOPC required consumers of grid square
information to validate it on each use. Many of them performed
only minimal validation.

Add a `Locator` object for representing a grid square. Only valid
grid squares are accepted as input.

Previous versions of ardopcf would transmit an encoded "`No GS`"
byte sequence when the TNC's `GRIDSQUARE` was unset. This was
probably an effort to avoid sending long runs of zeros, which in
some modulation schemes tend to cause sync drift. The `Locator`
object retains this behavior.

In [1], the no-grid byte sequence was inadvertently changed.
Previous versions of the compressor did not uppercase characters,
and the result of packing "`No GS`" was a corrupt byte sequence.

* `Locator` preserves the **new** behavior by sending the correct
  representation for "`NO GS`" (uppercase)

* `Locator` recognizes either byte sequence as an unpopulated
  grid square and will accept it without error.

* It is unknown how other implementations handle either byte
  sequence, which is not part of the spec.

The existing encoding scheme for grid squares is preserved. In
theory, with a resolution of 43200×43200 grid squares, it would be
possible to fit the square and resolution into just 35 bits
instead of the 48 it currently uses.

[1]: https://github.com/pflarue/ardop/commit/9d50e2a9ddc554e5a61f39f0aef59051563adb45