soundinput: replace callsign strings with smart objects
Replace all string-based callsigns in `SoundInput.c` with the new
`StationId` smart object. The ConReq, Ping, and IDFrame decoders
now take and populate `StationId` instead of string buffers.

Both the ConReq and Ping decoders communicate the decode to
higher-level logic (i.e., ARQ) via string buffers of the form
"`CALLER TARGET`." While we continue to populate this string,
we will eventually replace this with the new globals:

* `LastDecodedStationCaller`
* `LastDecodedStationTarget`

which hold the last successfully-decoded station IDs. We take
care to update these consistently and to zero them out with each
new frame.

The `lastGoodID` string appears to exist only to feed a UI
element. It is refactored out.