Remove all special handling of CQ as StationId.
Prior code relating to CQ was of limited utility and use.  Current
tests for valid callsigns allow any 2-character callsign.  So, CQ
can still be used if desired.  A user should not use CQ for MYCALL
or MYAUX, but using CQ is no worse than using any other illegal
callsign.  So, special tests to prevent this are not needed.  No
code exists or probably has existed to allow appropriate automatic
response to a Connect Request to CQ.  See recent discussion on users
group:
https://ardop.groups.io/g/users/topic/use_of_cq_with_ardop/108851759