log,common: add --syslog option on Linux
Add `--syslog` CLI option to send console logs to syslog(3)
instead of stderr. This new behavior is only available and
documented on Linux.

The syslog adapter is based on the example in zf_log [1]. Instead
of adding a NUL string terminator, we specify a printf(3)-style
*precision*. This keeps the text string compatible with the other
output mechanisms, which use that character for newline.

[1]: lib/zf_log/examples/custom_output.c