Apply a temporary fix for compatibility with beta firmware for the QRPLabs QMX
The Beta firmware 1_01_003 for the QRPLabs QMX and QMX+ has a known issue causing
PTT to disengage in USB mode if there is a pause in USB audio being sent.  This
minor change provides a workaround to avoid doing that.  Since I believe that the
code without this workaround will produce better results for some other radios, I
expect to revert this change when the QMX firmware is fixed.

However, I'm including this fix in the ardopcf develop branch for now since I'm
using the QMX for some on the air testing, and I think that with the new
capabilities provided by the beta firmware, the QMX and QMX+ are very well suited
to use ardopcf for low powered HF digital communications.